+++ hazelcast/src/hazelcast/client/serialization/pimpl/Packet.cpp
@@ -169,7 +169,7 @@ namespace hazelcast {
                         setPersistStatus(PERSIST_VERSION);
                         if (VERSION != version) {
                             std::stringstream stringstream;
-                            stringstream << "Packet versions are not matching! This -> " << VERSION << ", Incoming -> " << version;
+                            stringstream << "Packet versions are not matching! This -> " << (int)VERSION << ", Incoming -> " << (int)version;
                             throw exception::IllegalArgumentException("Packet::readVersion", stringstream.str());
                         }
                     }
